LRAC-12490 Fix individualSegmentId casting to Long

+++ modules/dxp/apps/osb/osb-asah/osb-asah-batch-curator/src/main/java/com/liferay/osb/asah/batch/curator/bot/nanite/DeleteIndividualSegmentTasksNanite.java
@@ -43,7 +43,7 @@ public class DeleteIndividualSegmentTasksNanite extends BaseNanite {
 
 		List<Long> individualSegmentIds = ListUtil.map(
 			jsonArray.toList(),
-			individualSegmentId -> (Long)individualSegmentId);
+			individualSegmentId -> Long.valueOf((int)individualSegmentId));
 
 		_bqIndividualInterestScoreDog.deleteBQIndividualInterestScores(
 			ListUtil.map(individualSegmentIds, String::valueOf));
